simplify storage
